+++ test/support/test-providers.ts
@@ -6,6 +6,7 @@ import * as Redacted from "effect/Redacted";
 import * as Provider from "alchemy/Provider";
 import { Bucket, BucketProvider } from "../../src/Bucket.ts";
 import { Container, ContainerProvider } from "../../src/Container.ts";
+import { ContainerImage, ContainerImageProvider } from "../../src/ContainerImage.ts";
 import { ScalewayCredentials } from "../../src/Credentials.ts";
 import { DatabaseInstance, DatabaseInstanceProvider } from "../../src/DatabaseInstance.ts";
 import { Domain, DomainProvider } from "../../src/Domain.ts";
@@ -46,6 +47,7 @@ export const testProviders = (options: { project?: ProjectRef } = {}) =>
       Namespace,
       Project,
       Container,
+      ContainerImage,
       Trigger,
       Domain,
       DnsZone,
@@ -70,6 +72,7 @@ export const testProviders = (options: { project?: ProjectRef } = {}) =>
         NamespaceProvider(),
         ProjectProvider(),
         ContainerProvider(),
+        ContainerImageProvider(),
         TriggerProvider(),
         DomainProvider(),
         DnsZoneProvider(),